Allow RMG to recognize "R_nondelocalized"
The MRH R_nondelocalized is now one of the allowed functional groups in the
RMG database, similar to Cs, Cbf, and R!H.  Furthermore, there are hard-coded
rules to recognize if R_nondelocalized is a parent of another node.

NOTE: R_nondelocalized is meant to represent any element that only has single
bonds attached to it.  This commit is the first in a series of commits MRH
plans to make in order to make RMG robust to heteroatom and aromatic
chemistry.