Updated RMG_database: Removed most hard-coded functional group references
MRH has updated most (not all) of the hard-coded functional groups (e.g. Cs, Cd,
Ct, etc.) in the RMG_database.  The *_Dictionary.txt files in the thermo_groups
folder have been updated as much as possible, as well as the */dictionary.txt
files in the kinetics_groups folder.

Right now, the "R_aromatic" functional group is in place.  However, MRH
believes this can be replaced with R_min2valency (since the "B" benzene bond
will distinguish if the node is aromatic or not).

This is only the database update.  The source code still needs to be resolved.