Updating how RMG handles max number of elements per species
Rather than require the user to specify the following to change the
maximum number of elements allowed per species:

MaxCarbonNumberPerSpecies: 5
MaxOyxgenNumberPerSpecies: 3
etc.

A user now specifies the more robust syntax:

MaxElementNumberPerSpecies C 5
MaxElementNumberPerSpecies O 3

This allows the same function as before, while allowing RMG to be robust
in which elements it may place a maximum limit on.  The only restriction
is that the element of interest be in the Element.txt file in the main
RMG_database directory.