Fixing bugs in calculateAtomSymmetryNumber()
1) Typo in list of arcs to compare (MRH had the 3rd and 4th arc twice,
when he meant to compare the 2nd and 4th arc)

2) Forgot to include howManyTrues case of 0 or 1 when checking atom with
four connections (before, RMG would throw an error)

3) MRH misinterpreted what this function was supposed to return (and I
was in fact combining the Atom and Axis Symmetry Numbers together).  If
the connections to the atom are not all equal (i.e. all single bonds),
this function should return an atom symmetry number of 1, which it now
does.