New attributes for ChemElement class
The attributes of each ChemElement are now stored and read in from the
Elements.txt file.  The function readListOfElements() reads in the file
line-by-line, creates a new ChemElement, and then adds that ChemElement
to the ChemElementDictionary, using the symbol as the key.

The ChemElement constructor has been updated accordingly.  New attributes
have been added to the ChemElement class.  The valency attribute is now
of type int (instead of double).