REBASE INTERACTIVE: This commit should go w/commence primary ignition
The new "initialization" function that should be called at the
beginning of any main class within the RMG default package has been
moved to the RMG class.  This commmit should be paired with it, as the
reading of forbidden structures (and MRH's new user-specified functional
group wildcards) was moved from this class.