Updating how RMG calculates the number of heavy atoms per molecule
Before, RMG would count the number of carbon, oxygen, sulfur, and silicon
atoms and assign that to the number of heavy atoms.

Now, RMG iterates over all elements in the Elements.txt file and determines
how many atoms are present in the molecule.  The number of hydrogen atoms
are then subtracted from this number, yielding the number of heavy atoms.