Store information regarding group matching (thermo, frequency)
When matching a group in the thermo_groups or frequency_groups library,
store the information when the ChemGraph.  This is particularly useful
for MRH when comparing RMG w/ and w/o hardcoded functional groups.

NOTE: MRH already pushed this to master, so when he merges this branch,
there will be a conflict.