Remove some of MRH's old hardcoded functional groups
At one point, MRH thought of having the hard-coded functional groups
R_aromatic, R_nonaromatic, and R_delocalized.  The first two are not
necessary because MRH decided to keep the "B" bond notation, so the
bond will determine if the atoms are aromatic or not; the last is not
necessary because MRH now uses the lumped bonding notation {#,{D,T,B}} to
represent that the atom of interest is delocalized, by being adjacent
to a non-single bond.