Allowing a set {} of FreeElectrons for any ChemGraph definition
Before: RMG expected to find a single place holder (e.g. "0", "1", "2S",
"2T", etc.) in the FreeElectron field of any ChemGraph.

However, in MRH's approach of making the software and database robust,
he required the use of sets of FreeElectrons (e.g. {0,1,2,2S,2T}).

Now: RMG will properly parse a set of FreeElectrons.  In the event the
user has supplied a set of FreeElectrons, a list of "Atom"s or "FGAtom"s
will be formed and returned from the readChemNodeElement() function.