Add robust code to check if a node is a previously-hardcoded FG
A few functions in the software ask "is this node a Cdd" (where Cdd had
the structure R1=C=R2) or similar.  In MRH's new implementation, RMG
does not know what a Cdd, Cs, etc. is.

This commit adds a function to robustly search for these old
hardcoded functional groups (w/o having to look for "Cdd").  However,
this function assumes all bonds to the node of interest have the same
strength.  So, searching for the old "Cs" or "Cdd" is OK; looking for
"Cd", "CO", or "Ct" would not work with this function.