Replacing "is this a Cs" with new function call
Instead of asking if a node is a "Cs", the question now becomes if the
node has exactly 4 connections, each having a bond strength of one?  When
MRH added silicon to RMG, assuming similar chemistry to carbon, this is
one spot he missed.  This more robust model solves the problem.