Replacing "Is this a Os or Ss" question with new implementation
Instead of asking "is this node a Os or Ss", the question now becomes
"does this node have exactly two connections, each with a bond order of
one?"

NOTE #1: MRH incorrectly put "Sis" in this function instead of "Ss" (since
it was assumed sulfur chemistry will mirror oxygen).

NOTE #2: MRH is assuming that an optical isomer exists for a S-S bond
(similar to the O-O bond).