Adding checks to ensure all molecule obey the new "Elements.txt" rules
The valency is no longer defined in the RMG software (i.e. a carbon can
have only 4 bonds).  This information (in the form of "maxNeighbors",
"maxHNeighbors", and "maxConnections") is now read in from the Elements.txt
file.

To ensure that all nodes of every species proposed by RMG follows these
rules, the new "valenciesOK" function has been added.  If any node in a
given species fails one of these checks, the species will not be made.