Updating the ChemElement class make() function
Before, RMG would look up the passed-in String in the ChemElementDictionary.
If not found, RMG would make a new ChemElement, put it in the Dictionary,
and proceed.

Now, RMG reads in and makes all ChemElements at the beginning of the
program (these are read in from the Elements.txt file).  If RMG now cannot
find the passed-in String in the ChemElementDictionary, the program will
fail.