Updating the "R 0 {#,X}" nodes with multiple FreeElectron field
When MRH updated the database by expanding the definitions in the
dictionary.txt files, he inadvertently made the definitions too specific.
For example, the Cs group was:

Cs
1 Cs 0

It then became:

Cs
1 C 0 {2,S} {3,S} {4,S} {5,S}
2 R 0 {1,S}
3 R 0 {1,S}
4 R 0 {1,S}
5 R 0 {1,S}

However, the "0" in the FreeElectron field was too-specific.  This really
should be "any value".  Thus, MRH changed the "0" to "{0,1,2,2S,2T,3}" to
keep with the old definition.